+++ src/Command/TestCommand.php
@@ -10,7 +10,7 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-#[AsCommand(name: 'app:test')]
+#[AsCommand(name: 'app:test', description: 'Internal test command - do not use in production.', hidden: true)]
 class TestCommand extends Command
 {
     public function __construct(