+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add command descriptions.
+### Fixed
+- Fix phpstan, psalm and code style problems.
 
 ## 0.0.15 - 2023-07-01
 ### Added